fix(tests): consolidate def-contract + wheel-ships tests into test__image_source_build.py
PS-204 §2 orphan-test-file: every test file must mirror exactly one
src file. Two of the new tests added in the previous commit had no
matching src file:

  * tests/scitex_agent_container/containers/test_def_install_from_bundled_source.py
    — the mirror dir src/scitex_agent_container/containers/ exists
    but holds only .def files, which the audit treats as "empty"
    (Python sources only).
  * tests/scitex_agent_container/cli_pkg/test_wheel_ships_bundled_files.py
    — no wheel_ships_bundled_files.py source file exists.

Both test contracts the staging helper (_image_source_build.py)
depends on, so they fit naturally there. Moved them into
tests/scitex_agent_container/cli_pkg/test__image_source_build.py
and deleted the orphan files. Test behaviour is unchanged: the
wheel-ships tests still skip when pip is unavailable; the .def
contract tests still parametrize over all three shipped .defs.

This fixes the pytest-matrix failure on py3.11/3.12/3.13 that
blocked PR #255 merge.
